Rework tutorial based on feedback
